Fix unsigned vs signed warnings spotted by clang on OpenBSD
(which is more picky than others)

+++ pdns/test-iputils_hh.cc
@@ -259,13 +259,13 @@ BOOST_AUTO_TEST_CASE(test_NetmaskGroup) {
   {
     NetmaskGroup ng;
     BOOST_CHECK_EQUAL(ng.empty(), true);
-    BOOST_CHECK_EQUAL(ng.size(), 0);
+    BOOST_CHECK_EQUAL(ng.size(), 0U);
     ng.addMask("10.0.1.0");
     BOOST_CHECK_EQUAL(ng.empty(), false);
-    BOOST_CHECK_EQUAL(ng.size(), 1);
+    BOOST_CHECK_EQUAL(ng.size(), 1U);
     BOOST_CHECK(ng.match(ComboAddress("10.0.1.0")));
     ng.toMasks("127.0.0.0/8, 10.0.0.0/24");
-    BOOST_CHECK_EQUAL(ng.size(), 3);
+    BOOST_CHECK_EQUAL(ng.size(), 3U);
     BOOST_CHECK(ng.match(ComboAddress("127.0.0.1")));
     BOOST_CHECK(ng.match(ComboAddress("10.0.0.3")));
     BOOST_CHECK(ng.match(ComboAddress("10.0.1.0")));
@@ -273,26 +273,26 @@ BOOST_AUTO_TEST_CASE(test_NetmaskGroup) {
     BOOST_CHECK(!ng.match(ComboAddress("10.0.1.1")));
     BOOST_CHECK(!ng.match(ComboAddress("::1")));
     ng.addMask("::1");
-    BOOST_CHECK_EQUAL(ng.size(), 4);
+    BOOST_CHECK_EQUAL(ng.size(), 4U);
     BOOST_CHECK(ng.match(ComboAddress("::1")));
     BOOST_CHECK(!ng.match(ComboAddress("::2")));
     ng.addMask("fe80::/16");
-    BOOST_CHECK_EQUAL(ng.size(), 5);
+    BOOST_CHECK_EQUAL(ng.size(), 5U);
     BOOST_CHECK(ng.match(ComboAddress("fe80::1")));
     BOOST_CHECK(!ng.match(ComboAddress("fe81::1")));
     BOOST_CHECK_EQUAL(NMGOutputToSorted(ng.toString()), NMGOutputToSorted("10.0.1.0/32, 127.0.0.0/8, 10.0.0.0/24, ::1/128, fe80::/16"));
 
     /* negative entries using the explicit flag */
     ng.addMask("172.16.0.0/16", true);
-    BOOST_CHECK_EQUAL(ng.size(), 6);
+    BOOST_CHECK_EQUAL(ng.size(), 6U);
     BOOST_CHECK(ng.match(ComboAddress("172.16.1.1")));
     BOOST_CHECK(ng.match(ComboAddress("172.16.4.50")));
     ng.addMask("172.16.4.0/24", false);
-    BOOST_CHECK_EQUAL(ng.size(), 7);
+    BOOST_CHECK_EQUAL(ng.size(), 7U);
     BOOST_CHECK(ng.match(ComboAddress("172.16.1.1")));
     BOOST_CHECK(!ng.match(ComboAddress("172.16.4.50")));
     ng.addMask("fe80::/24", false);
-    BOOST_CHECK_EQUAL(ng.size(), 8);
+    BOOST_CHECK_EQUAL(ng.size(), 8U);
     BOOST_CHECK(!ng.match(ComboAddress("fe80::1")));
     BOOST_CHECK(!ng.match(ComboAddress("fe81::1")));
     /* not in fe80::/24 but in fe80::/16, should match */
@@ -301,12 +301,12 @@ BOOST_AUTO_TEST_CASE(test_NetmaskGroup) {
     /* negative entries using '!' */
     BOOST_CHECK(ng.match(ComboAddress("172.16.10.80")));
     ng.addMask("!172.16.10.0/24");
-    BOOST_CHECK_EQUAL(ng.size(), 9);
+    BOOST_CHECK_EQUAL(ng.size(), 9U);
     BOOST_CHECK(!ng.match(ComboAddress("172.16.10.80")));
     ng.addMask("2001:db8::/32");
-    BOOST_CHECK_EQUAL(ng.size(), 10);
+    BOOST_CHECK_EQUAL(ng.size(), 10U);
     ng.addMask("!2001:db8::/64");
-    BOOST_CHECK_EQUAL(ng.size(), 11);
+    BOOST_CHECK_EQUAL(ng.size(), 11U);
     BOOST_CHECK(!ng.match(ComboAddress("2001:db8::1")));
     /* not in 2001:db8::/64 but in 2001:db8::/32, should match */
     BOOST_CHECK(ng.match(ComboAddress("2001:db8:1::1")));
@@ -318,15 +318,15 @@ BOOST_AUTO_TEST_CASE(test_NetmaskGroup) {
     /* this time using Netmask objects instead of strings */
     NetmaskGroup ng;
     BOOST_CHECK_EQUAL(ng.empty(), true);
-    BOOST_CHECK_EQUAL(ng.size(), 0);
+    BOOST_CHECK_EQUAL(ng.size(), 0U);
     ng.addMask(Netmask("10.0.1.0"));
     BOOST_CHECK_EQUAL(ng.empty(), false);
-    BOOST_CHECK_EQUAL(ng.size(), 1);
+    BOOST_CHECK_EQUAL(ng.size(), 1U);
     BOOST_CHECK(ng.match(ComboAddress("10.0.1.0")));
     ng.addMask(Netmask("127.0.0.0/8"));
-    BOOST_CHECK_EQUAL(ng.size(), 2);
+    BOOST_CHECK_EQUAL(ng.size(), 2U);
     ng.addMask(Netmask("10.0.0.0/24"));
-    BOOST_CHECK_EQUAL(ng.size(), 3);
+    BOOST_CHECK_EQUAL(ng.size(), 3U);
     BOOST_CHECK(ng.match(ComboAddress("127.0.0.1")));
     BOOST_CHECK(ng.match(ComboAddress("10.0.0.3")));
     BOOST_CHECK(ng.match(ComboAddress("10.0.1.0")));
@@ -334,26 +334,26 @@ BOOST_AUTO_TEST_CASE(test_NetmaskGroup) {
     BOOST_CHECK(!ng.match(ComboAddress("10.0.1.1")));
     BOOST_CHECK(!ng.match(ComboAddress("::1")));
     ng.addMask(Netmask("::1"));
-    BOOST_CHECK_EQUAL(ng.size(), 4);
+    BOOST_CHECK_EQUAL(ng.size(), 4U);
     BOOST_CHECK(ng.match(ComboAddress("::1")));
     BOOST_CHECK(!ng.match(ComboAddress("::2")));
     ng.addMask(Netmask("fe80::/16"));
-    BOOST_CHECK_EQUAL(ng.size(), 5);
+    BOOST_CHECK_EQUAL(ng.size(), 5U);
     BOOST_CHECK(ng.match(ComboAddress("fe80::1")));
     BOOST_CHECK(!ng.match(ComboAddress("fe81::1")));
     BOOST_CHECK_EQUAL(NMGOutputToSorted(ng.toString()), NMGOutputToSorted("10.0.1.0/32, 127.0.0.0/8, 10.0.0.0/24, ::1/128, fe80::/16"));
 
     /* negative entries using the explicit flag */
     ng.addMask(Netmask("172.16.0.0/16"), true);
-    BOOST_CHECK_EQUAL(ng.size(), 6);
+    BOOST_CHECK_EQUAL(ng.size(), 6U);
     BOOST_CHECK(ng.match(ComboAddress("172.16.1.1")));
     BOOST_CHECK(ng.match(ComboAddress("172.16.4.50")));
     ng.addMask(Netmask("172.16.4.0/24"), false);
-    BOOST_CHECK_EQUAL(ng.size(), 7);
+    BOOST_CHECK_EQUAL(ng.size(), 7U);
     BOOST_CHECK(ng.match(ComboAddress("172.16.1.1")));
     BOOST_CHECK(!ng.match(ComboAddress("172.16.4.50")));
     ng.addMask("fe80::/24", false);
-    BOOST_CHECK_EQUAL(ng.size(), 8);
+    BOOST_CHECK_EQUAL(ng.size(), 8U);
     BOOST_CHECK(!ng.match(ComboAddress("fe80::1")));
     BOOST_CHECK(!ng.match(ComboAddress("fe81::1")));
     /* not in fe80::/24 but in fe80::/16, should match */
@@ -366,14 +366,14 @@ BOOST_AUTO_TEST_CASE(test_NetmaskGroup) {
 BOOST_AUTO_TEST_CASE(test_NetmaskTree) {
   NetmaskTree<int> nmt;
   BOOST_CHECK_EQUAL(nmt.empty(), true);
-  BOOST_CHECK_EQUAL(nmt.size(), 0);
+  BOOST_CHECK_EQUAL(nmt.size(), 0U);
   nmt.insert(Netmask("130.161.252.0/24")).second=0;
   BOOST_CHECK_EQUAL(nmt.empty(), false);
-  BOOST_CHECK_EQUAL(nmt.size(), 1);
+  BOOST_CHECK_EQUAL(nmt.size(), 1U);
   nmt.insert(Netmask("130.161.0.0/16")).second=1;
-  BOOST_CHECK_EQUAL(nmt.size(), 2);
+  BOOST_CHECK_EQUAL(nmt.size(), 2U);
   nmt.insert(Netmask("130.0.0.0/8")).second=2;
-  BOOST_CHECK_EQUAL(nmt.size(), 3);
+  BOOST_CHECK_EQUAL(nmt.size(), 3U);
 
   BOOST_CHECK_EQUAL(nmt.lookup(ComboAddress("213.244.168.210")), (void*)0);
   auto found=nmt.lookup(ComboAddress("130.161.252.29"));
@@ -392,27 +392,27 @@ BOOST_AUTO_TEST_CASE(test_NetmaskTree) {
   BOOST_CHECK_EQUAL(found->second, 2);
 
   nmt.insert(Netmask("0.0.0.0/0")).second=3;
-  BOOST_CHECK_EQUAL(nmt.size(), 4);
+  BOOST_CHECK_EQUAL(nmt.size(), 4U);
   nmt.insert(Netmask("0.0.0.0/7")).second=4;
-  BOOST_CHECK_EQUAL(nmt.size(), 5);
+  BOOST_CHECK_EQUAL(nmt.size(), 5U);
   nmt.insert(Netmask("0.0.0.0/15")).second=5;
-  BOOST_CHECK_EQUAL(nmt.size(), 6);
+  BOOST_CHECK_EQUAL(nmt.size(), 6U);
   BOOST_CHECK_EQUAL(nmt.lookup(Netmask("0.0.0.0/0"))->second, 3);
   BOOST_CHECK_EQUAL(nmt.lookup(Netmask("0.0.0.0/7"))->second, 4);
   BOOST_CHECK_EQUAL(nmt.lookup(Netmask("0.0.0.0/15"))->second, 5);
 
   nmt.clear();
   BOOST_CHECK_EQUAL(nmt.empty(), true);
-  BOOST_CHECK_EQUAL(nmt.size(), 0);
+  BOOST_CHECK_EQUAL(nmt.size(), 0U);
   BOOST_CHECK(!nmt.lookup(ComboAddress("130.161.180.1")));
 
   nmt.insert(Netmask("::1")).second=1;
   BOOST_CHECK_EQUAL(nmt.empty(), false);
-  BOOST_CHECK_EQUAL(nmt.size(), 1);
+  BOOST_CHECK_EQUAL(nmt.size(), 1U);
   nmt.insert(Netmask("::/0")).second=0;
-  BOOST_CHECK_EQUAL(nmt.size(), 2);
+  BOOST_CHECK_EQUAL(nmt.size(), 2U);
   nmt.insert(Netmask("fe80::/16")).second=2;
-  BOOST_CHECK_EQUAL(nmt.size(), 3);
+  BOOST_CHECK_EQUAL(nmt.size(), 3U);
   BOOST_CHECK_EQUAL(nmt.lookup(ComboAddress("130.161.253.255")), (void*)0);
   BOOST_CHECK_EQUAL(nmt.lookup(ComboAddress("::2"))->second, 0);
   BOOST_CHECK_EQUAL(nmt.lookup(ComboAddress("::ffff"))->second, 0);
@@ -423,19 +423,19 @@ BOOST_AUTO_TEST_CASE(test_NetmaskTree) {
 BOOST_AUTO_TEST_CASE(test_single) {
   NetmaskTree<bool> nmt;
   BOOST_CHECK_EQUAL(nmt.empty(), true);
-  BOOST_CHECK_EQUAL(nmt.size(), 0);
+  BOOST_CHECK_EQUAL(nmt.size(), 0U);
   nmt.insert(Netmask("127.0.0.0/8")).second=1;
   BOOST_CHECK_EQUAL(nmt.empty(), false);
-  BOOST_CHECK_EQUAL(nmt.size(), 1);
+  BOOST_CHECK_EQUAL(nmt.size(), 1U);
   BOOST_CHECK_EQUAL(nmt.lookup(ComboAddress("127.0.0.1"))->second, 1);
 }
 
 BOOST_AUTO_TEST_CASE(test_scale) {
   string start="192.168.";
   NetmaskTree<int> works;
-  BOOST_CHECK_EQUAL(works.size(), 0);
-  for(int i=0; i < 256; ++i) {
-    for(int j=0; j < 256; ++j) {
+  BOOST_CHECK_EQUAL(works.size(), 0U);
+  for(size_t i=0; i < 256; ++i) {
+    for(size_t j=0; j < 256; ++j) {
       works.insert(Netmask(start+std::to_string(i)+"."+std::to_string(j))).second=i*j;
       BOOST_CHECK_EQUAL(works.size(), i*256 + j + 1);
     }
@@ -455,8 +455,8 @@ BOOST_AUTO_TEST_CASE(test_scale) {
   }
 
   start="2000:123:";
-  for(int i=0; i < 256; ++i) {
-    for(int j=0; j < 256; ++j) {
+  for(size_t i=0; i < 256; ++i) {
+    for(size_t j=0; j < 256; ++j) {
       works.insert(Netmask(start+std::to_string(i)+":"+std::to_string(j)+"::/64")).second=i*j;
       BOOST_CHECK_EQUAL(works.size(), (256*256) + i*256 + j + 1);
     }
@@ -480,7 +480,7 @@ BOOST_AUTO_TEST_CASE(test_removal) {
   std::string prefix = "192.";
   NetmaskTree<int> nmt;
   BOOST_CHECK(nmt.empty());
-  BOOST_CHECK_EQUAL(nmt.size(), 0);
+  BOOST_CHECK_EQUAL(nmt.size(), 0U);
 
   size_t count = 0;
   for(unsigned int i = 0; i < 256; ++i) {
@@ -620,8 +620,8 @@ BOOST_AUTO_TEST_CASE(test_iterator) {
 
   // check set equality using iterators
   BOOST_CHECK_EQUAL(masks_set1.size(), masks_set2.size());
-  BOOST_CHECK_EQUAL(std::distance(masks_set1.begin(), masks_set1.end()),
-                    std::distance(masks_set2.begin(), masks_set2.end()));
+  BOOST_CHECK_EQUAL((size_t)std::distance(masks_set1.begin(), masks_set1.end()),
+                    (size_t)std::distance(masks_set2.begin(), masks_set2.end()));
   for (auto entry: masks_set1) {
     Netmask mask = entry.first.getNormalized();
 
@@ -640,8 +640,8 @@ BOOST_AUTO_TEST_CASE(test_iterator) {
 
   // check set equality using iterators
   BOOST_CHECK_EQUAL(masks_set1_cp1.size(), masks_set2.size());
-  BOOST_CHECK_EQUAL(std::distance(masks_set1_cp1.begin(), masks_set1_cp1.end()),
-                    std::distance(masks_set2.begin(), masks_set2.end()));
+  BOOST_CHECK_EQUAL((size_t)std::distance(masks_set1_cp1.begin(), masks_set1_cp1.end()),
+                    (size_t)std::distance(masks_set2.begin(), masks_set2.end()));
   for (auto entry: masks_set1_cp1) {
     Netmask mask = entry.first.getNormalized();
 
@@ -660,8 +660,8 @@ BOOST_AUTO_TEST_CASE(test_iterator) {
 
   // check set equality using iterators
   BOOST_CHECK_EQUAL(masks_set1_cp2.size(), masks_set2.size());
-  BOOST_CHECK_EQUAL(std::distance(masks_set1_cp2.begin(), masks_set1_cp2.end()),
-                    std::distance(masks_set2.begin(), masks_set2.end()));
+  BOOST_CHECK_EQUAL((size_t)std::distance(masks_set1_cp2.begin(), masks_set1_cp2.end()),
+                    (size_t)std::distance(masks_set2.begin(), masks_set2.end()));
   for (auto entry: masks_set1_cp2) {
     Netmask mask = entry.first.getNormalized();
 
@@ -681,8 +681,8 @@ BOOST_AUTO_TEST_CASE(test_iterator) {
 
   // check set equality using iterators
   BOOST_CHECK_EQUAL(masks_set1_cp3.size(), masks_set2.size());
-  BOOST_CHECK_EQUAL(std::distance(masks_set1_cp3.begin(), masks_set1_cp3.end()),
-                    std::distance(masks_set2.begin(), masks_set2.end()));
+  BOOST_CHECK_EQUAL((size_t)std::distance(masks_set1_cp3.begin(), masks_set1_cp3.end()),
+                    (size_t)std::distance(masks_set2.begin(), masks_set2.end()));
   for (auto entry: masks_set1_cp3) {
     Netmask mask = entry.first.getNormalized();
 
@@ -776,8 +776,8 @@ BOOST_AUTO_TEST_CASE(test_iterator) {
 
   // check set equality using iterators
   BOOST_CHECK_EQUAL(masks_set1_cp5.size(), masks_set2.size());
-  BOOST_CHECK_EQUAL(std::distance(masks_set1_cp5.begin(), masks_set1_cp5.end()),
-                    std::distance(masks_set2.begin(), masks_set2.end()));
+  BOOST_CHECK_EQUAL((size_t)std::distance(masks_set1_cp5.begin(), masks_set1_cp5.end()),
+                    (size_t)std::distance(masks_set2.begin(), masks_set2.end()));
   for (auto entry: masks_set1_cp5) {
     Netmask mask = entry.first.getNormalized();
